fix: change guardianId type Long to Integer

+++ backend/src/main/java/com/storydream/backend/domain/guardian/controller/AuthController.java
@@ -32,7 +32,7 @@ public class AuthController {
     }
 
     @GetMapping("/guardians/me")
-    public ResponseEntity<ApiResponse<GuardianResponse>> getMe(@AuthenticationPrincipal Long guardianId) {
+    public ResponseEntity<ApiResponse<GuardianResponse>> getMe(@AuthenticationPrincipal Integer guardianId) {
         return ResponseEntity.ok(ApiResponse.success(authService.getMe(guardianId)));
     }
 }
